requirements.txt
discord.py==1.5.1
jishaku
psycopg2
Pillow